Update languages.py - breaking condition for .q
Modifying comment for q language files to protect against edge case. 

line 20 - '.q': {'language': 'q', 'comment': '/'}, -> '.q': {'language': 'q', 'comment': '//'},

Edge case: Empty lines in Markdown were causing the entire rest of the script to be commented out. In line with q syntax, "/" on a single line comments remainder of script. 

Proposed change: Amendment of existing comment "/" to double comment "//" also serves as a comment for individual lines, but without that side effect of commenting the full script in the event of an empty line.